+++ src/wizard.rs
@@ -1,0 +1,55 @@
+use amethyst::{
+    prelude::*,
+    core::transform::Transform,
+    renderer::{Camera}
+};
+
+use amethyst::winit::*;
+
+pub struct WizardState;
+pub const ARENA_HEIGHT: f32 = 100.0;
+pub const ARENA_WIDTH: f32 = 100.0;
+
+impl SimpleState for WizardState {
+    fn on_start(&mut self, _data: StateData<'_, GameData<'_, '_>>) {
+        println!("Hello world");
+        initialize_camera(_data.world);
+    }
+
+    fn on_stop(&mut self, _data: StateData<'_, GameData<'_, '_>>) {
+        println!("Goodbye world");
+    }
+
+    // fn update(&mut self, _: &mut StateData<'_, GameData<'_, '_>>) -> SimpleTrans {
+    //     println!("hello from amethyst");
+    //     Trans::Quit
+    // }
+
+    fn handle_event(&mut self, _data: StateData<'_, GameData<'_, '_>>, event: StateEvent) -> SimpleTrans {
+        if let StateEvent::Window(event) = &event {
+            match event {
+                Event::WindowEvent { event , .. } => match event {
+                    WindowEvent::KeyboardInput {
+                        input: KeyboardInput { virtual_keycode: Some(VirtualKeyCode::Escape), .. }, ..
+                    } |
+                    WindowEvent::CloseRequested => Trans::Quit,
+                    _ => Trans::None
+                },
+                _ => Trans::None
+            }
+        } else {
+            Trans::None
+        }
+    }
+}
+
+fn initialize_camera(world: &mut World) {
+    let mut transform = Transform::default();
+    transform.set_translation_xyz(ARENA_WIDTH * 0.5, ARENA_HEIGHT * 0.5, 1.0);
+
+    world
+        .create_entity()
+        .with(Camera::standard_2d(ARENA_WIDTH, ARENA_HEIGHT))
+        .with(transform)
+        .build();
+}
